chore(tests): test setup command
